requirements.txt
Django==2.2
pytz==2018.9
sqlparse==0.3.0
django-heroku==0.3.1
python-decouple
coverage
psycopg2-binary
gunicorn==19.8.1
arxiv
python-dateutil